backend/auth_services/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.13
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosmtplib==3.0.2
anyio==4.8.0
asgiref==3.8.1
attrs==25.2.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Deprecated==1.2.18
distlib==0.3.9
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
filelock==3.17.0
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
multidict==6.1.0
phonenumbers==8.13.55
pkce==1.0.3
platformdirs==4.3.6
propcache==0.3.0
pycparser==2.22
pycryptodome==3.20.0
PyJWT==2.10.1
pyotp==2.9.0
python-dateutil==2.9.0.post0
requests==2.32.3
requests-file==2.1.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
supertokens_python==0.29.0
tldextract==5.1.3
twilio==9.5.0
typing_extensions==4.12.2
urllib3==2.3.0
virtualenv==20.29.3
wrapt==1.17.2
yarl==1.18.3
